Move node handlers into their own directory
