+++ packages/arbor-store/src/index.ts
@@ -1,11 +1,11 @@
 export { Arbor, ImmutableArbor } from "./Arbor"
-export { ArrayNodeHandler } from "./ArrayNodeHandler"
 export * from "./decorators"
 export * from "./errors"
 export * from "./guards"
-export { MapNodeHandler } from "./MapNodeHandler"
+export { ArrayNodeHandler } from "./handlers/ArrayNodeHandler"
+export { MapNodeHandler } from "./handlers/MapNodeHandler"
+export { NodeHandler } from "./handlers/NodeHandler"
 export * from "./MutationEngine"
-export { NodeHandler } from "./NodeHandler"
 export { Path } from "./Path"
 export * from "./ScopedStore"
 export { Subscribers } from "./Subscribers"
